feat: added ts in project

+++ src/server.ts
@@ -1,29 +1,29 @@
-import http from 'http';
+import { createServer, IncomingMessage, ServerResponse } from 'http';
 import {
   getAllUsersController,
   getUserByIdController,
   createUserController,
   updateUserController,
   deleteUserController
-} from './controllers/usersControllers.js';
+} from './controllers/usersControllers';
 import dotenv from 'dotenv';
 
 dotenv.config();
 
-http.createServer((req, res) => {
+createServer((req: IncomingMessage, res: ServerResponse) => {
   if(req.url === '/api/users' && req.method === 'GET') {
     
     getAllUsersController(req, res);
-  } else if (req.url.match(/\/api\/users\/([0-9a-z]+)/) && req.method === 'GET') {
+  } else if (req.url && req.url.match(/\/api\/users\/([0-9a-z]+)/) && req.method === 'GET') {
 
     getUserByIdController(req, res);
   } else if(req.url === '/api/users' && req.method === 'POST') {
 
     createUserController(req, res);
-  } else if(req.url.match(/\/api\/users\/([0-9a-z]+)/) && req.method === 'PUT') {
+  } else if(req.url && req.url.match(/\/api\/users\/([0-9a-z]+)/) && req.method === 'PUT') {
 
     updateUserController(req, res);
-  } else if(req.url.match(/\/api\/users\/([0-9a-z]+)/) && req.method === 'DELETE') {
+  } else if(req.url && req.url.match(/\/api\/users\/([0-9a-z]+)/) && req.method === 'DELETE') {
 
     deleteUserController(req, res);
   } else {
